move aggregate functions to separate file
